# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- Partners ----------
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, titre, description, icon, sortOrder]
      properties:
        id:
          type: integer
        titre:
          type: string
        description:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    CreateServiceBody:
      type: object
      required: [titre, description, icon]
      properties:
        titre:
          type: string
        description:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer
    UpdateServiceBody:
      type: object
      properties:
        titre:
          type: string
        description:
          type: string
        icon:
          type: string
        sortOrder:
          type: integer

    # ---------- Partners ----------
...