# lib/api-zod/src/generated/types/updateServiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServiceBody {
  titre?: string;
  description?: string;
  icon?: string;
  sortOrder?: number;
}